requirements/base.txt

celery==4.0.2

Django==1.11.1

django-filter==1.0.2
django-rest-swagger==2.1.2
djangorestframework==3.6.3
whitenoise==3.3.0
dj-database-url==0.4.2
dj-email-url==0.0.10
django-twilio
coverage==4.4.1
django-coverage-plugin==1.4.2
nltk==3.2.5
numpy==1.14.1
simplejson==3.13.2
django-bootstrap-form==3.3
django-bootstrap3 == 9.1
social-auth-app-django == 2.1.0
redis == 2.10.6




